When changing roles of a user (as an admin), I made the roles mutually exclusive which cleaned up a decent amount of code.
Removed the option of creating user roles.

Cleaned up the Manage view.

Added some input validation on the forms and moved the error display to the top of the page and made it much more obvious.

TODO: I plan to revisit the layout of the Manage Roles to make it a bit more fluid and easier to use.