Compressed the Manage Users view into a single username/role box. This forced a bit of a rework in the controller.
Added the ability to delete a user but it has not been fully tested. The display and action of the button seem to function but the controller actually deleting the user has not been used quite yet. I plan to test that next time.

TODO: Tidy up code and document it a bit more.