Remade the database so the correct table information was present.
Finished the delete user method so users are removed from logins/roles/user. All information is gone.

I changed the way roles worked a little bit. Admins can no longer be removed from the admin role or change to a different role. Once an admin - always an admin. Instead of creating a "super-admin", to prevent system lockout, I decided that if a user is added to the admin class then they should always be that class. The only way to remove the user from that class is to delete them from the database completely. This may change in the future if I can find an easier way to mitigate the possibility of system lock-out.

I removed a lot of required fields from the genome model but I am still having problems with the wizard beliving that JumpLength is a required field when is definitely is not. That is currently unresolved.