Fixed the issue with the model not validating properly. MVC implicitly requires value types to have values so setting them to nullable serves two purposes:
(1) Solves the model validation problem.
(2) Allows me to catch the cases in which they didn't input a value and then set a default in its place before saving to the database so there will never been a null value in the DB.

Removed the .edmx in the web config since that is no longer required.