Added a initial empty string value for the data source to prevent page load problems when referencing an uninitialized object.
Added dynamically added textboxes for n-number of URLs. There is a bug that I am currently working on however. When both functions appendURLBox and concatURLs are in the same code block, they conflict and neither will run. But I depend on the variable i which is incremented by appendURLBox to reference (and grab the value) in concatURLs.