Added error handling so we can not only stop the connection if there is an error but also return the actual error to the user for further debugging and logging. Will eventually output this information to a log file as well.
Additionally, if there are any errors the model is NOT added to the db but is returned to the user for further inspection to make sure the information is good.

The path is still tentative. It will be passed from the ConfigBuilder which will contain the URL to the data to which I will append the name of the particular file for the wget.