Cleaned the code up as bit and added in some error checking for specific situations.
TODO: To fix the issue with permissions and checking on the status of a job, we need to change the directory to /tmp/Genome and chmod the folder for everyone to view and work with so we can update the information. Hopefully this is temporary. But that ought to work.