Cleaned a bit of code up.
Changed the path to be in the config folder prior to calling the assemble script or adding it to the scheduler with qsub. However, I think that will be a problem since I ran into problems adding it to the queue when I wasn't on the login node. So we might need to address that by changing our pathing around a bit.