Added a couple minor bug fixes.
Added some custom CSS that represents the feedback boxes that appear throughout the program. Instead of hardcoding it, might as well throw it into some custom css. To that point, I changed the way the output is shown across the website to be more uniform and mellow. It was bright yellow and red in some places.

Created a function to get a list of all the users so you know exactly who to reference when changing a user's roles or deleting them.

Added a stop-check to make sure that there isn't a single admin left when attempting to delete a user. To prevent lockout, I procluded the possibility of deleting a user if they are the LAST admin (aka 1 user with admin roles left).