Finished the majority of the logic for the updateall method as well as changed the way we handle static paths in the code.
Since we end up referencing a lot of the same data (at least 2 or 3 times) across the project, I setup a class that will get all of the data paths if it is given the uuid of the particular job.

I also changed the way I handle the step storage and reference as well. I passed that off to a file which can be accessed across the project. Static storage in a single class might have been ok, but there is no reason to not store it in another class in case I want to do something with it later.