Finished the updateall method but a lot of it needs to be methodized so that the statuses will update correctly.
As it stands now, if a job is deemed to be "completed" (success OR failure), it will carry out doing the compression/sftp connection/uploading of the data. Assuming the worst case, that could be a couple hundred gbs. That could take awhile and we would want to move onto other jobs AND give those intermediate updates to the user as to what their job is doing.

Unfortunately, there is no way for me to currently check the logic of this aside from black boxing each command. I can't see how they work in consort until we get a viable testbed setup to run a full job from start to finish.