Methodized the job upload portion of the update.
I changed the way the update actually runs. I renamed it and gave it a jobsToUpload which is passed by ref and allows me to specify whether the particular job being passed into the method is in need of upload to the FTP in the calling location. Handy since I need to know whether or not to call UploadData.